p-interface: Change name for Pinocchio interface crate
#### Problem

As part of https://github.com/orgs/anza-xyz/projects/27, we want to
separate interfaces from programs for Agave.

The token repo already has `spl-token-interface`, but it's
Pinocchio-only, which means it'll be currently difficult to start using
in Agave.

#### Summary of changes

Change the crate name to `pinocchio-token-interface` and directory name
to `p-interface` so that we can add `interface` in a future PR.