deps: Bump sdk to v3, decouple from token program
#### Problem

The interface crate is still on the v2 sdk crates, but the v3 crates are
ready.

#### Summary of changes

Bump the sdk crates to v3 and decouple the program from the repo version
of the interface crate. The only change is the `to_str` signature on the
error.

While doing this, I noticed that the interface crate wasn't actually
part of the workspace, so I added it in.