Postconditions check if value is `Err` before calling `unwrap` (#30)
Previously I had missed checks that `Result` types were not `Err` before
calling unwrap, now there are explicit branches and appropriate error
codes that reflect P-Token's behaviour in this case. I have not been
able to test this concretely with the tests, we are going to have to
roll with the proofs.

I also fixed:
- Outdated commands in the documentation for running the proofs
- A copy paste bug in `TransferChecked`

I intend to run the proofs overnight and update the documents that store
the info in the morning

Closes https://github.com/runtimeverification/solana-token/issues/23