Added "multisig" feature flag and branched proofs accordingly (#32)
An owner / authority can be either an `Account` or a `Multisig` and so
we must conditionally call the appropriate cheatcode based on which case
we would like to prove. Through conditional compilation we can do this
without having to add other flags into the harnesses, and we only
compile the relevant code for what we are intending to prove - this
might be better for artefact size but I haven't done any diagnostics on
this. This flag is disabled by default, but I updated the documentation
so there is a record of it existing.

Currently running `./run_verification.sh -a` to get an idea on how this
improves (hopeful) the verification.

Another idea I had was adding conditional panics in the multisig branch
of the `Validate Owner` logic:
```
#[cfg(not(feature="multisig"))]
unreachable!()
```
Since we shouldn't be in that branch if we are a multisig anyway, this
is more defensive than an empty block.

There also might be other areas where a similar idea would benefit us,
but I just wanted to try this idea out for now.