Move pinocchio crates into subdirectory
#### Problem

The Makefile model uses a slightly hacky way of distinguishing
subdirectories, where anything before the first `-` is taken to be a
directory name. This is a problem for `p-token` and `p-interface`, since
the Makefile will look in `p/token` and `p/interface`.

#### Summary of changes

Move p-token into pinocchio/program and p-interface into
pinocchio/interface.

Update everything else as needed.