program: Upgrade to v3 crates
#### Problem

The interface crate has been on v3 for some time, but the program and
tests have been lagging behind.

#### Summary of changes

Upgrade everything to new CLI, crates, etc, and fix everything that
comes up.

The biggest change is that it isn't possible to disable direct mapping,
because it was removed from the feature set in v3.0.