CI: Require TOML to be formatted
tombi is a pretty good TOML formatter/LSP with support for TOML in
general, and Rust Cargo.toml in particular.

Using a unified format is nice.

In particular, tombi sorts not just properties within a section, but
also sections and tables.  Apparently some TOML parsers may fail if
tables are not correctly sorted.